refactor update wallet reducer function to properly use promises
